+++ jwst/assign_wcs/util.py
@@ -18,6 +18,7 @@ from gwcs import WCS
 from gwcs.wcstools import wcs_from_fiducial, grid_from_bounding_box
 from gwcs import utils as gwutils
 from stdatamodels import DataModel
+from stpipe.exceptions import StpipeExitException
 
 from . import pointing
 from ..lib.catalog_utils import SkyObject
@@ -40,7 +41,7 @@ class MSAFileError(Exception):
         super(MSAFileError, self).__init__(message)
 
 
-class NoDataOnDetectorError(Exception):
+class NoDataOnDetectorError(StpipeExitException):
     """WCS solution indicates no data on detector
 
     When WCS solutions are available, the solutions indicate that no data
@@ -56,7 +57,9 @@ class NoDataOnDetectorError(Exception):
     def __init__(self, message=None):
         if message is None:
             message = 'WCS solution indicate that no science is in the data.'
-        super(NoDataOnDetectorError, self).__init__(message)
+        # The first argument instructs stpipe CLI tools to exit with status
+        # 64 when this exception is raised.
+        super().__init__(64, message)
 
 
 def _domain_to_bounding_box(domain):