+++ jwst/associations/lib/product_utils.py
@@ -3,7 +3,9 @@
 from collections import defaultdict, Counter
 import copy
 import logging
+import warnings
 
+from .. import config
 from .diff import compare_product_membership
 
 logger = logging.getLogger(__name__)
@@ -15,12 +17,12 @@ def sort_by_candidate(asns):
 
     Parameters
     ----------
-    asns: [Association[,...]]
+    asns : [Association[,...]]
         List of associations
 
     Returns
     -------
-    sorted_by_candidate: [Associations[,...]]
+    sorted_by_candidate : [Associations[,...]]
         New list of the associations sorted.
 
     Notes
@@ -39,11 +41,11 @@ def get_product_names(asns):
 
     Parameters
     ----------
-    asns: [`Association`[, ...]]
+    asns : [`Association`[, ...]]
 
     Returns
     -------
-    product_names, duplicates: set(str[, ...]), [str[,...]]
+    product_names, duplicates : set(str[, ...]), [str[,...]]
         2-tuple consisting of the set of product names and the list of duplicates.
     """
     product_names = [
@@ -67,19 +69,23 @@ def get_product_names(asns):
 def prune_duplicate_associations(asns):
     """Remove duplicate associations in favor of lower level versions
 
-    For Level 3 associations, multiple associations with the same membership, but different
-    levels, can be created. Remove duplicate associations of higher level.
+    Main use case: For Level 3 associations, multiple associations with the
+    same membership, but different levels, can be created. Remove duplicate
+    associations of higher level.
 
-    The assumption is that there is only one product per association, before merging.
+    The assumption is that there is only one product per association, before
+    merging.
 
     Parameters
     ----------
-    asns: [Association[,...]]
+    asns : [Association[,...]]
         Associations to prune
 
     Returns
-    pruned: [Association[,...]]
+    -------
+    pruned : [Association[,...]]
         Pruned list of associations
+
     """
     ordered_asns = sort_by_candidate(asns)
     pruned = list()
@@ -105,11 +111,8 @@ def prune_duplicate_associations(asns):
 def prune_duplicate_products(asns):
     """Remove duplicate products in favor of higher level versions
 
-    For Level 2 associations, since the products are always just the input
-    exposures, different candidates can be created for each exposure. Remove
-    those associations of lesser candidates.
-
-    The assumption is that there is only one product per association, before merging
+    The assumption is that there is only one product per association, before
+    merging
 
     Parameters
     ----------
@@ -119,10 +122,17 @@ def prune_duplicate_products(asns):
     Returns
     pruned: [Association[,...]]
         Pruned list of associations
+
     """
     product_names, dups = get_product_names(asns)
     if not dups:
         return asns
+
+    warnings.warn(f'Duplicate associations exist: {dups}', RuntimeWarning)
+    if config.DEBUG:
+        warnings.warn('Duplicate associations will have "dupXXX" prepended to their names, where "XXX" is a 3-digit sequence.')
+    else:
+        warnings.warn('Duplicates will be removed, leaving only one of each.', RuntimeWarning)
 
     pruned = copy.copy(asns)
     to_prune = defaultdict(list)
@@ -131,9 +141,14 @@ def prune_duplicate_products(asns):
         if product_name in dups:
             to_prune[product_name].append(asn)
 
+    dup_count = 0
     for product_name, asns_to_prune in to_prune.items():
         asns_to_prune = sort_by_candidate(asns_to_prune)
         for asn in asns_to_prune[1:]:
-            pruned.remove(asn)
+            if config.DEBUG:
+                dup_count += 1
+                asn.asn_name = f'dup{dup_count:03d}_{asn.asn_name}'
+            else:
+                pruned.remove(asn)
 
     return pruned