+++ jwst/associations/lib/rules_level2b.py
@@ -423,6 +423,45 @@ class Asn_Lv2MIRLRSFixedSlitNod(
 
 
 @RegistryMarker.rule
+class Asn_Lv2NRSLAMPImage(
+        AsnMixin_Lv2Image,
+        AsnMixin_Lv2Special,
+        DMSLevel2bBase
+):
+    """Level2b NIRSpec image Lamp Calibrations Association
+
+    Characteristics:
+        - Association type: ``image2``
+        - Pipeline: ``calwebb_image2``
+        - Image-based calibration exposures
+        - Single science exposure
+    """
+
+    def __init__(self, *args, **kwargs):
+        # Setup constraints
+        self.constraints = Constraint([
+            Constraint_Base(),
+            Constraint_Single_Science(self.has_science),
+            DMSAttrConstraint(
+                name='exp_type',
+                sources=['exp_type'],
+                value='nrs_lamp'
+            ),
+            DMSAttrConstraint(
+                sources=['grating'],
+                value='mirror'
+            ),
+            DMSAttrConstraint(
+                sources=['opmode'],
+                value='image',
+                required=False
+            ),
+        ])
+
+        super(Asn_Lv2NRSLAMPImage, self).__init__(*args, **kwargs)
+
+
+@RegistryMarker.rule
 class Asn_Lv2NRSLAMPSpectral(
         AsnMixin_Lv2Special,
         DMSLevel2bBase
@@ -598,7 +637,7 @@ class Asn_Lv2WFSS(
             for idx in sorted(direct_idxs, reverse=True)
         ))
 
-        # Add the Level3 catalog and direct image members
+        # Add the Level3 catalog, direct image, and segmentation map members
         lv3_direct_image_root = DMS_Level3_Base._dms_product_name(self)
         members.append(
             Member({
@@ -610,6 +649,12 @@ class Asn_Lv2WFSS(
             Member({
                 'expname': lv3_direct_image_root + '_cat.ecsv',
                 'exptype': 'sourcecat'
+            })
+        )
+        members.append(
+            Member({
+                'expname': lv3_direct_image_root + '_segm.fits',
+                'exptype': 'segmap'
             })
         )