+++ jwst/associations/lib/rules_level3_base.py
@@ -38,9 +38,9 @@ from jwst.associations.lib.dms_base import (
     IMAGE2_NONSCIENCE_EXP_TYPES,
     SPEC2_SCIENCE_EXP_TYPES,
 )
-from jwst.associations.lib.format_template import FormatTemplate
+from stpipe.format_template import FormatTemplate
 from jwst.associations.lib.member import Member
-from jwst.associations.lib.product_utils import prune_duplicate_associations
+from jwst.associations.lib.product_utils import prune_duplicate_associations, prune_duplicate_products
 
 __all__ = [
     'ASN_SCHEMA',
@@ -600,6 +600,7 @@ class Utility():
                 finalized_asns.append(asn)
 
         lv3_asns = prune_duplicate_associations(lv3_asns)
+        lv3_asns = prune_duplicate_products(lv3_asns)
 
         # Ensure sequencing is correct.
         Utility.resequence(lv3_asns)