+++ jwst/stpipe/tests/test_pipeline.py
@@ -4,8 +4,10 @@ import sys
 import numpy as np
 import pytest
 
+from stpipe import crds_client
+
 from jwst import datamodels
-from jwst.stpipe import Step, Pipeline, crds_client
+from jwst.stpipe import Step, Pipeline
 
 from .steps import PipeWithReference, StepWithReference
 
@@ -104,7 +106,7 @@ class MyPipeline(Pipeline):
 
 def test_pipeline(_jail):
     pipeline_fn = join(dirname(__file__), 'steps', 'python_pipeline.cfg')
-    pipe = Step.from_config_file(pipeline_fn)
+    pipe = Pipeline.from_config_file(pipeline_fn)
     pipe.output_filename = "output.fits"
 
     assert pipe.flat_field.threshold == 42.0
@@ -139,7 +141,7 @@ def test_prefetch(_jail, monkeypatch):
     class MockGetRef:
         called = False
 
-        def mock(self, dataset_model, reference_file_types, observatory=None):
+        def mock(self, parameters, reference_file_types, observatory):
             if 'flat' in reference_file_types:
                 self.called = True
             result = {