+++ CHANGES.rst
@@ -1,6 +1,14 @@
 0.18.4 (unreleased)
 ===================
 
+associations
+------------
+
+- Warn about duplicate product names and do not write duplicate associations [#5721]
+
+- Added new Lvl2 rule, Asn_Lv2NRSLAMPImage, to run Image2 pipeline for NRSLAMP
+  exposures with OPMODE=image [#5740]
+
 cube_build
 ----------
 
@@ -15,21 +23,27 @@ datamodels
 - Added extr_x and extr_y to multispec datamodel. These values are center
   of extraction region for IFU data [#5685]
 
+- Added segmentation map output file name to core schema keywords, under
+  keyword 'SEGMFILE' [#5730]
+
+- Added '1LOS' to PATTTYPE enum list in core.schema datamodel [#5728]
+
+- Added 'IMAGE' to OPMODE enum list [#5745]
+
 extract_1d
 ----------
 
 - Adding writing SRCTYPE, EXTR_X, and EXTR_Y to extracted spec for IFU data [#5685]
+
 - Only update the output x1d data using the PRIMARY input data. Prevents SCI data in x1d data [#5694]
+
+- Fixed bug in background region fitting for image columns/rows that have zero weight
+  for all pixels [#5696]
 
 lib
 ---
 
 - Make EngDB_Value public for JSDP use [#5669]
-
-ramp_fitting
-------------
-
-- Update documentation to define optimal weighting algorithm [#5682]
 
 - Update code in ``set_velocity_aberration.py`` functions based on Colin Cox
   suggestions: simplify DVA scale computation and improve apparent ``RA`` and
@@ -41,12 +55,36 @@ ramp_fitting
 
 - Make get_wcs_values_from_siaf public for JSDP use [#5669]
 
+ramp_fitting
+------------
+
+- Fixed bug in handling NGROUPS=2 exposures for pixels that saturate in group 2.
+  Proper slope, err, and other quantities are now computed from the good data
+  in group 1. [#5700]
+
+- Update documentation to define optimal weighting algorithm [#5682]
+
+source_catalog
+--------------
+
+- Added the segmentation map as an output data file, with
+  suffix "segm". [#5730]
+
 srctype
 -------
 
 - Changed default SRCTYPE for non-primary NIRSpec slits in a FIXEDSLIT
   exposure to 'EXTENDED' rather than 'POINT' [#5671]
 
+- Changed logic for handling NIRSpec MOS exposures to blank out the "global"
+  value of SRCTYPE, to ensure that only the individual slit-specific values
+  of SRCTYPE get used downstream. [#5754]
+
+stpipe
+------
+
+- Make jwst.stpipe independent of the rest of the jwst package and move
+  core code to spacetelescope/stpipe. [#5695, #5720]
 
 0.18.3 (2021-01-25)
 ===================
@@ -198,6 +236,8 @@ assign_wcs
 
 associations
 ------------
+
+- Asn_Lv2WFSS: Add segmentation map exposure to Level2 WFSS associations [#5532]
 
 - Add new dither keyword subpxpts to constraints [#5525]